+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.4
-* the target directory is created automatically, if it does not exist.
+* the target directory is created automatically, if it does not exist;
+* added `ASSETS` constant.
 
 ### 1.1.3
 * added `AssetNotFoundException` class. This exception is thrown when you